requirements.txt
langchain
openai
faiss-cpu
pypdf
python-dotenv
streamlit
tiktoken
pdfminer.six
urllib3
